refactor: split MCP tool contract from wiring in core
McpToolDefinition/McpToolInput move to core/presentation/mcp_tool/mcp_tool.ts,
paralleling core/presentation/controller/controller.ts. The SDK wiring
(registerMcpTool) stays in infra, renamed to mcp_tool_adapter.ts to avoid two
mcp_tool.ts files in different layers. Tools still living in their old
location already import the type from its new home.
