+++ src/core/infra/mcp/tools/cancel_stay.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import type { StayDi } from "../../../../booking/infra/di/stay_di";
-import type { McpToolDefinition } from "../mcp_tool";
+import type { McpToolDefinition } from "../../../presentation/mcp_tool/mcp_tool";
 
 const inputSchema = {
   stay_id: z
