+++ src/core/infra/mcp/tools/list_property_settings.ts
@@ -5,7 +5,7 @@ import {
   MAX_LIMIT,
 } from "../../../application/dto/pagination";
 import type { PropertyManagementDi } from "../../../../property_management/infra/di/property_management_di";
-import type { McpToolDefinition } from "../mcp_tool";
+import type { McpToolDefinition } from "../../../presentation/mcp_tool/mcp_tool";
 
 const inputSchema = {
   property_id: z
