+++ src/core/infra/mcp/tools/update_property_setting.ts
@@ -4,7 +4,7 @@ import {
   boundedJsonValue,
   settingTypeSchema,
 } from "../../../domain/value_object/setting_value";
-import type { McpToolDefinition } from "../mcp_tool";
+import type { McpToolDefinition } from "../../../presentation/mcp_tool/mcp_tool";
 
 const inputSchema = {
   property_id: z
